Modularize initialization
