+++ espnet/nets/pytorch_backend/e2e_asr.py
@@ -26,6 +26,8 @@ from chainer import reporter
 from espnet.nets.asr_interface import ASRInterface
 from espnet.nets.e2e_asr_common import label_smoothing_dist
 from espnet.nets.pytorch_backend.ctc import ctc_for
+from espnet.nets.pytorch_backend.initialization import lecun_normal_init_parameters
+from espnet.nets.pytorch_backend.initialization import set_forget_bias_to_one
 from espnet.nets.pytorch_backend.nets_utils import pad_list
 from espnet.nets.pytorch_backend.nets_utils import to_device
 from espnet.nets.pytorch_backend.nets_utils import to_torch_tensor
@@ -41,6 +43,7 @@ class Reporter(chainer.Chain):
     """A chainer reporter wrapper."""
 
     def report(self, loss_ctc, loss_att, acc, cer_ctc, cer, wer, mtl_loss):
+        """Report at every step."""
         reporter.report({'loss_ctc': loss_ctc}, self)
         reporter.report({'loss_att': loss_att}, self)
         reporter.report({'acc': acc}, self)
@@ -59,6 +62,7 @@ class E2E(ASRInterface, torch.nn.Module):
     :param Namespace args: argument Namespace containing options
 
     """
+
     @staticmethod
     def add_arguments(parser):
         """Add arguments."""
@@ -231,33 +235,6 @@ class E2E(ASRInterface, torch.nn.Module):
         - EmbedID.W ~ Normal(0, 1)
         - LSTM.upward.b[forget_gate_range] = 1 (but not used in NStepLSTM)
         """
-
-        def lecun_normal_init_parameters(module):
-            for p in module.parameters():
-                data = p.data
-                if data.dim() == 1:
-                    # bias
-                    data.zero_()
-                elif data.dim() == 2:
-                    # linear weight
-                    n = data.size(1)
-                    stdv = 1. / math.sqrt(n)
-                    data.normal_(0, stdv)
-                elif data.dim() in (3, 4):
-                    # conv weight
-                    n = data.size(1)
-                    for k in data.size()[2:]:
-                        n *= k
-                    stdv = 1. / math.sqrt(n)
-                    data.normal_(0, stdv)
-                else:
-                    raise NotImplementedError
-
-        def set_forget_bias_to_one(bias):
-            n = bias.size(0)
-            start, end = n // 4, n // 2
-            bias.data[start:end].fill_(1.)
-
         lecun_normal_init_parameters(self)
         # exceptions
         # embed weight ~ Normal(0, 1)
@@ -387,7 +364,12 @@ class E2E(ASRInterface, torch.nn.Module):
         return dict(decoder=self.dec, ctc=CTCPrefixScorer(self.ctc, self.eos))
 
     def encode(self, x):
-        """Encode acoustic features."""
+        """Encode acoustic features.
+
+        :param ndarray x: input acoustic feature (T, D)
+        :return: encoder outputs
+        :rtype: torch.Tensor
+        """
         self.eval()
         ilens = [x.shape[0]]
 
@@ -410,7 +392,7 @@ class E2E(ASRInterface, torch.nn.Module):
         return hs.squeeze(0)
 
     def recognize(self, x, recog_args, char_list, rnnlm=None):
-        """E2E beam search
+        """E2E beam search.
 
         :param ndarray x: input acoustic feature (T, D)
         :param Namespace recog_args: argument Namespace containing options
@@ -432,7 +414,7 @@ class E2E(ASRInterface, torch.nn.Module):
         return y
 
     def recognize_batch(self, xs, recog_args, char_list, rnnlm=None):
-        """E2E beam search
+        """E2E beam search.
 
         :param list xs: list of input acoustic feature arrays [(T_1, D), (T_2, D), ...]
         :param Namespace recog_args: argument Namespace containing options
@@ -478,7 +460,7 @@ class E2E(ASRInterface, torch.nn.Module):
         return y
 
     def enhance(self, xs):
-        """Forwarding only the frontend stage
+        """Forwarding only the frontend stage.
 
         :param ndarray xs: input acoustic feature (T, C, F)
         """
