+++ espnet/nets/pytorch_backend/initialization.py
@@ -1,0 +1,29 @@
+import math
+
+
+def lecun_normal_init_parameters(module):
+    for p in module.parameters():
+        data = p.data
+        if data.dim() == 1:
+            # bias
+            data.zero_()
+        elif data.dim() == 2:
+            # linear weight
+            n = data.size(1)
+            stdv = 1. / math.sqrt(n)
+            data.normal_(0, stdv)
+        elif data.dim() in (3, 4):
+            # conv weight
+            n = data.size(1)
+            for k in data.size()[2:]:
+                n *= k
+            stdv = 1. / math.sqrt(n)
+            data.normal_(0, stdv)
+        else:
+            raise NotImplementedError
+
+
+def set_forget_bias_to_one(bias):
+    n = bias.size(0)
+    start, end = n // 4, n // 2
+    bias.data[start:end].fill_(1.)
